Revert "Some eatable fixes"
This reverts commit 52da2f0f9a8a47449a1a945890dc62fe0998a19c.

+++ src/xrGame/ui/UICellItem.cpp
@@ -307,8 +307,8 @@ void CUICellItem::UpdateConditionProgressBar()
 			CEatableItem* eitm = smart_cast<CEatableItem*>(itm);
 			if (eitm)
 			{
-				u8 max_uses = eitm->GetMaxUses();
-				u8 remaining_uses = eitm->GetRemainingUses();
+				u16 max_uses = eitm->GetMaxUses();
+				u16 remaining_uses = eitm->GetRemainingUses();
 
 				if (max_uses < 8)
 				{
